replace FroseAiAggregator with FedAvgAggregator and restructure aggregator module
