+++ froseai/aggregator/__init__.py
@@ -1,0 +1,5 @@
+from .agg_mix import FedAvgAggregator
+
+__all__ = [
+    "FedAvgAggregator"
+]
