+++ froseai/flow/__init__.py
@@ -1,2 +1,1 @@
-from .aggregator import FedAvgAggregator as FroseAiAggregator
 from .optimizer import FedAvg as FroseAiOptimizer
