+++ froseai/pb/README.md
@@ -1,0 +1,6 @@
+# FroseAi protobuf
+### Build
+
+```shell
+protoc --python_out=. ./froseai.proto
+```
